Additional Command Line Argument Features
Added command line arguments for batch executions. Now executable packed with all of the written code and doesn't need compiler directives to use some of the options. Other CA rules are given as an option and an alternative for key scrambling, though not recommended. Details are found in README.txt.